requirements.txt
aiogram==3.7.0
aiohttp==3.9.5
asyncpg==0.29.0
cloudscraper==1.2.71
python-dotenv==1.0.1
pyrogram==2.0.106
tgcrypto==1.2.5
